requirements.txt
scipy
# CRITICAL: Adjust this line to match your GPU cluster's CUDA version.
cupy-cuda12x
